+++ tests/unit/with_db/test_tools_backup.py
@@ -1,20 +1,14 @@
 """Tests for backup tool."""
 import os
-from pathlib import Path
-import pytest
-
-CURR_DIR = Path(os.path.dirname(os.path.realpath(__file__)))
-METADATA_FOLDER = CURR_DIR / ".." / "data" / "metadata"
-MANIFEST = CURR_DIR / ".." / "data" / "manifest.txt"
+from filecmp import cmpfiles
 
 
-@pytest.mark.skip(reason="These tests require movai-core-enterprise")
 class TestToolsBackup:
-    def test_import_manifest(self, global_db):
+    def test_import_manifest(self, global_db, metadata_folder, manifest_file):
         from dal.tools.backup import Importer
 
         tool = Importer(
-            METADATA_FOLDER,
+            metadata_folder,
             force=True,
             dry=False,
             debug=False,
@@ -22,11 +16,11 @@ class TestToolsBackup:
             clean_old_data=True,
         )
 
-        objects = tool.read_manifest(MANIFEST)
+        objects = tool.read_manifest(manifest_file)
 
         tool.run(objects)
 
-    def test_export_manifest(self, global_db, tmp_path):
+    def test_export_manifest(self, global_db, manifest_file, tmp_path):
         from dal.tools.backup import Exporter
 
         tool = Exporter(
@@ -35,6 +29,85 @@ class TestToolsBackup:
             recursive=False,
         )
 
-        objects = tool.read_manifest(MANIFEST)
+        objects = tool.read_manifest(manifest_file)
 
         tool.run(objects)
+
+    def test_relative_import(self, global_db, metadata_folder, manifest_file):
+        from dal.tools.backup import Importer
+
+        METADATA_FOLDER_RELATIVE = metadata_folder.relative_to(os.getcwd())
+        MANIFEST_RELATIVE = manifest_file.relative_to(os.getcwd())
+
+        tool = Importer(
+            METADATA_FOLDER_RELATIVE,
+            force=True,
+            dry=False,
+            debug=False,
+            recursive=False,
+            clean_old_data=True,
+        )
+
+        objects = tool.read_manifest(MANIFEST_RELATIVE)
+
+        tool.run(objects)
+
+    def test_import_export_translation(self, global_db, metadata_folder, manifest_file, tmp_path):
+        """Test translation import and export."""
+        from dal.tools.backup import Importer, Exporter
+
+        importer = Importer(
+            metadata_folder,
+            force=True,
+            dry=False,
+            debug=False,
+            recursive=False,
+            clean_old_data=True,
+        )
+
+        data = {"Translation": ["delete_me"]}
+
+        importer.run(data)
+
+        exporter = Exporter(
+            tmp_path,
+            debug=False,
+            recursive=False,
+        )
+
+        exporter.run(data)
+
+        to_check = [
+            "delete_me.json",
+            "delete_me_pt.po",
+            "delete_me_fr.po",
+        ]
+
+        equal, diff, err = cmpfiles(
+            metadata_folder / "Translation", tmp_path / "Translation", to_check
+        )
+        assert set(equal) == set(to_check)
+        assert not diff
+        assert not err
+
+    def test_import_invalid_data(
+        self, global_db, metadata_folder_invalid_data, manifest_file_invalid_data, capsys
+    ):
+        """Test import validates and reports invalid data."""
+        from dal.tools.backup import Importer
+
+        tool = Importer(
+            metadata_folder_invalid_data,
+            force=True,
+            dry=False,
+            debug=False,
+            recursive=False,
+            clean_old_data=True,
+        )
+
+        objects = tool.read_manifest(manifest_file_invalid_data)
+
+        tool.run(objects)
+
+        captured = capsys.readouterr()
+        assert "Failed to import, invalid schema for 'Translation:delete_me'" in captured.out